+++ source/core/transform/idwt_neon.cpp
@@ -111,8 +111,8 @@ auto idwt_irrev97_fixed_neon_hor_step3 = [](const int32_t init_pos, const int32_
 
 void idwt_1d_filtr_irrev97_fixed_neon(sprec_t *X, const int32_t left, const int32_t u_i0,
                                       const int32_t u_i1) {
-  const auto i0        = static_cast<const int32_t>(u_i0);
-  const auto i1        = static_cast<const int32_t>(u_i1);
+  const auto i0        = static_cast<int32_t>(u_i0);
+  const auto i1        = static_cast<int32_t>(u_i1);
   const int32_t start  = i0 / 2;
   const int32_t stop   = i1 / 2;
   const int32_t offset = left - i0 % 2;
@@ -137,8 +137,8 @@ void idwt_1d_filtr_irrev97_fixed_neon(sprec_t *X, const int32_t left, const int3
 // reversible IDWT
 void idwt_1d_filtr_rev53_fixed_neon(sprec_t *X, const int32_t left, const int32_t u_i0,
                                     const int32_t u_i1) {
-  const auto i0        = static_cast<const int32_t>(u_i0);
-  const auto i1        = static_cast<const int32_t>(u_i1);
+  const auto i0        = static_cast<int32_t>(u_i0);
+  const auto i1        = static_cast<int32_t>(u_i1);
   const int32_t start  = i0 / 2;
   const int32_t stop   = i1 / 2;
   const int32_t offset = left - i0 % 2;
@@ -370,7 +370,7 @@ void idwt_rev_ver_sr_fixed_neon(sprec_t *in, const int32_t u0, const int32_t u1,
       for (int32_t col = simdlen; col < u1 - u0; ++col) {
         int32_t sum = buf[n - 1][col];
         sum += buf[n + 1][col];
-        buf[n][col] -= ((sum + 2) >> 2);
+        buf[n][col] -= static_cast<int16_t>((sum + 2) >> 2);
       }
     }
     for (int32_t n = 0 + offset, i = start; i < stop; ++i, n += 2) {
@@ -384,7 +384,7 @@ void idwt_rev_ver_sr_fixed_neon(sprec_t *in, const int32_t u0, const int32_t u1,
       for (int32_t col = simdlen; col < u1 - u0; ++col) {
         int32_t sum = buf[n][col];
         sum += buf[n + 2][col];
-        buf[n + 1][col] += (sum >> 1);
+        buf[n + 1][col] += static_cast<int16_t>(sum >> 1);
       }
     }
